Add an option fatal log level behind a feature flag
Some environments require an additional fatal log level. So this
includes this additional log level with a cargo feature flag. The fatal
log level is the highest priority, one above the error level.
This change is fully backwards compatible. Users or crates, that do not
enable the 'fatal' feature will see no change in behavior or API.